Update UI.java
The saved and load game methods are a major hurdle. I think that one way to handle this issue is to place all of the boolean values of the other class into one class named SavedGame(). The issue is that if they are  private, it wont work. Maybe a protected and make SavedGame a superclass or subclass. 
To load, simply open the SavedGame and call the SavedGameloop() which will call all the other classes to retrieve the boolean values and position values of the older game.